+++ src/Trarizon.Library.Collections/HISTORY.md
@@ -3,3 +3,4 @@
 ## 1.0.1
 
 - Reimplement `PrefixTree(Dictionary)` with node, the original implementation is renamed to `ContiguousPrefixTree(Dicitionary)` and marked as experimental
+- Add `TraAlgorithm.LevenshteinDistance`